requirements.txt
django~=3.0.6
web3~=5.10.0
py-solc